EAMxx: move column ops to util lib
