+++ components/eamxx/src/dynamics/homme/eamxx_homme_process_interface.cpp
@@ -30,7 +30,7 @@
 #include "dynamics/homme/interface/eamxx_homme_interface.hpp"
 #include "physics/share/physics_constants.hpp"
 #include "share/physics/eamxx_common_physics_functions.hpp"
-#include "share/algorithm/eamxx_column_ops.hpp"
+#include "share/util/eamxx_column_ops.hpp"
 #include "share/property_checks/field_lower_bound_check.hpp"
 
 // Ekat includes
