+++ components/eamxx/src/physics/iop_forcing/eamxx_iop_forcing_process_interface.hpp
@@ -3,7 +3,7 @@
 
 #include "share/atm_process/atmosphere_process.hpp"
 #include "share/atm_process/ATMBufferManager.hpp"
-#include "share/algorithm/eamxx_column_ops.hpp"
+#include "share/util/eamxx_column_ops.hpp"
 
 #include "physics/share/physics_constants.hpp"
