+++ components/eamxx/src/physics/rrtmgp/eamxx_rrtmgp_process_interface.cpp
@@ -8,7 +8,7 @@
 #include "share/algorithm/eamxx_fv_phys_rrtmgp_active_gases_workaround.hpp"
 #include "share/property_checks/field_within_interval_check.hpp"
 #include "share/physics/eamxx_common_physics_functions.hpp"
-#include "share/algorithm/eamxx_column_ops.hpp"
+#include "share/util/eamxx_column_ops.hpp"
 
 #include <ekat_team_policy_utils.hpp>
 #include <ekat_assert.hpp>
