+++ components/eamxx/src/share/physics/eamxx_common_physics_functions_impl.hpp
@@ -2,7 +2,7 @@
 #define SCREAM_COMMON_PHYSICS_IMPL_HPP
 
 #include "physics/share/physics_constants.hpp"
-#include "share/algorithm/eamxx_column_ops.hpp"
+#include "share/util/eamxx_column_ops.hpp"
 
 namespace scream {
