+++ components/eamxx/src/share/tests/CMakeLists.txt
@@ -2,9 +2,6 @@
 if (NOT SCREAM_ONLY_GENERATE_BASELINES)
   include(ScreamUtils)
 
-  # Test column ops
-  CreateUnitTest(column_ops "column_ops.cpp")
-
   # Generate data for data interpolation test
   CreateUnitTest(data_interpolation_setup "data_interpolation_setup.cpp"
     LIBS scream_io
