+++ components/eamxx/src/share/util/eamxx_column_ops.hpp
